Ensure the result from the parse contains only natvie python types.
This specifically was an issue with pyats as it was returning a qdict.

This was found to be an issue when running with forks<hosts, as the controller
needed to resolve the parser specific types back to native types and caused
a hang on the second batch of hosts during the import.