network: netconf: support XML huge_tree for RPC calls
__rpc__ method of NetconfConnection employs "fromstring" function
to parse the XML response. If a node in the XML tree is too big,
the parser fails with:

  "lxml.etree.XMLSyntaxError: xmlSAX2Characters: huge text node"

This is actually a libxml2 security feature that can be optionally
tweaked with the "huge_tree" option of the XML parser.

Module junos_config of the junipernetworks.junos collection employs
NetconfConnection.__rpc__ to get the configuration from a device.
There can be a case when configuration size will exceed the XML
parser limit and the module will bail out.

To facilitate such cases it should be possible for the RPC method
caller to enable the "huge_tree" option, which defaults to false.
Let the method support that by passing an appropriate parser to
"fromstring" function.