Allow processing of a specific configuration file for the connection
Depending on the destination, libssh might need to be configured in a
way, that the algorithms and methods need to be adjusted. Instead of
adding each possible configuration option into the wrapper, including
a config file (as it exists for ordinary OpenSSH) is much easier. This
allows ansible to connect to devices with less well supported
algorithms.

This is a required step to fix connection issues with
ansible.netcommon.libssh. The called ansible-pylibssh need to
process the new optional argument config_file. But this is a separate
patch, which needs to be applied there.