Allow processing of a specific configuration file for libssh (#498)
* Allow processing of a specific configuration file for the connection

Depending on the destination, libssh might need to be configured in a
way, that the algorithms and methods need to be adjusted. Instead of
adding each possible configuration option into the wrapper, including
a config file (as it exists for ordinary OpenSSH) is much easier. This
allows ansible to connect to devices with less well supported
algorithms.

This is a required step to fix connection issues with
ansible.netcommon.libssh. The called ansible-pylibssh need to
process the new optional argument config_file. But this is a separate
patch, which needs to be applied there.

* Provide a changelog fragment

* Update phrasing

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: lutz <lutz@iks-service.de>
Co-authored-by: Kate Case <this.is@katherineca.se>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>